core: provide package_list() and utilize in issue-diff.py.
